api/requirements.txt
Django==3.0.2
djangorestframework==3.11.0
psycopg2-binary==2.8.4
python-dateutil==2.8.1
requests==2.23.0
django-filter==2.2.0
django-crontab==0.7.1